refactor(ffi): split witness boundary module
